Admin-Portal-KYB-App/ai-score-service-main/requirements.txt
scikit-learn==1.5.2
xgboost==2.1.3
pandas==2.2.3
numpy==2.0.2
scipy==1.14.1
python-dotenv==1.0.1
boto3==1.34.49
confluent-kafka==2.3.0
shap==0.46.0
pyarrow==18.1.0
fastparquet==2024.2.0
botocore==1.34.66
simplejson==3.19.3
polars==1.17.1
torch>=2.7.1-rc1
pytest==9.0.0